moving multi_(has/get)_attr from output utils to inputs.py (only used in inputs.py) and made them both private
