+++ src/pymatgen/io/jdftx/_output_utils.py
@@ -83,59 +83,6 @@ def read_outfile_slices(file_name: str) -> list[list[str]]:
     return texts
 
 
-def multi_hasattr(varbase: Any, varname: str):
-    """Check if object has an attribute (capable of nesting with . splits).
-
-    Check if object has an attribute (capable of nesting with . splits).
-
-    Parameters
-    ----------
-    varbase
-        Object to check.
-    varname
-        Attribute to check for.
-
-    Returns
-    -------
-    bool
-        Whether the object has the attribute.
-    """
-    varlist = varname.split(".")
-    for i, var in enumerate(varlist):
-        if i == len(varlist) - 1:
-            return hasattr(varbase, var)
-        if hasattr(varbase, var):
-            varbase = getattr(varbase, var)
-        else:
-            return False
-    return None
-
-
-def multi_getattr(varbase: Any, varname: str):
-    """Check if object has an attribute (capable of nesting with . splits).
-
-    Check if object has an attribute (capable of nesting with . splits).
-
-    Parameters
-    ----------
-    varbase
-        Object to check.
-    varname
-        Attribute to check for.
-
-    Returns
-    -------
-    Any
-        Attribute of the object.
-    """
-    if not multi_hasattr(varbase, varname):
-        raise AttributeError(f"{varbase} does not have attribute {varname}")
-    varlist = varname.split(".")
-    for var in varlist:
-        varbase = getattr(varbase, var)
-    return varbase
-
-
 def _brkt_list_of_3_to_nparray(line: str) -> np.ndarray:
     """Return 3x1 numpy array.
